install-firefox.sh: make install-firefox idempotent

+++ scripts/install-firefox.sh
@@ -1,14 +1,24 @@
 #!/bin/bash -x
 
 cd /tmp/
-FFTAR="$(pwd)/firefox.tar.bz2"
-FF="firefox"
-sudo wget 'https://download.mozilla.org/?product=firefox-latest&os=linux64&lang=en-US' -O "${FFTAR}" || exit ${LINENO}
+FFTAR="firefox.tar.bz2"
+FF=$(basename "${FFTAR}" .tar.bz2)
 
-cd /usr/local/
-sudo tar -axvf "${FFTAR}" || exit ${LINENO}
-cd "${FF}"
+if ! test -f ${FFTAR}; then
+    sudo wget 'https://download.mozilla.org/?product=firefox-latest&os=linux64&lang=en-US' -O "${FFTAR}" || exit ${LINENO}
+fi
 
-EXE="$(pwd)/firefox/firefox-bin"
+INSTALL_DIR=/usr/local/${FF}
+
+if ! test -d ${INSTALL_DIR}; then
+    sudo tar -C -axvf "${FFTAR}" || exit ${LINENO}
+fi
+
+cd  || exit ${LINENO}
+
+
+EXE="${INSTALL_DIR}/firefox-bin"
+test -f ${EXE} || exit ${LINENO}
+
 sudo update-alternatives --install /usr/local/bin/firefox firefox "${EXE}" 20
 sudo update-alternatives --set firefox "${EXE}"
